continue even if we can't determine interface name
Some boxes do not have way to map IP address to interface name, there is
no need to give up, as we already handle this case as long as we invent
some name to them.
In this instance, both are unknown/without priority, and hence old/first
IP will be kept in DB if it still answers, newer will only replace it,
if old is dead.